Use Long for chat_id on all methods
Fixes #395

+++ telegramium-core/src/main/scala/telegramium/bots/client/BanChatSenderChatReq.scala
@@ -7,4 +7,4 @@ import telegramium.bots.ChatId
   * @param senderChatId
   *   Unique identifier of the target sender chat
   */
-final case class BanChatSenderChatReq(chatId: ChatId, senderChatId: Int)
+final case class BanChatSenderChatReq(chatId: ChatId, senderChatId: Long)
