+++ telegramium-core/src/main/scala/telegramium/bots/client/GetChatMenuButtonReq.scala
@@ -3,4 +3,4 @@ package telegramium.bots.client
 /** @param chatId
   *   Unique identifier for the target private chat. If not specified, default bot's menu button will be returned
   */
-final case class GetChatMenuButtonReq(chatId: Option[Int] = Option.empty)
+final case class GetChatMenuButtonReq(chatId: Option[Long] = Option.empty)
